+++ js/apps/system/_admin/aardvark/APP/react/src/views/views/editView/useFetchViewProperties.tsx
